# internal/api/openapi.yaml
openapi: 3.1.0
info:
  title: ACS Storage REST API
  version: "1"
  description: |
    Management and integration API for ACS Storage. The S3-compatible API is
    served separately (default port 9000) and follows the AWS S3 specification.

    **Authentication** — use a session cookie (web panel) or an access key:
    `Authorization: Bearer <ACCESS_KEY_ID>:<SECRET>` or HTTP Basic auth with the
    key ID as user name and the secret as password.

    **Errors** use RFC 9457 `application/problem+json`. POST and PATCH requests
    must send `Content-Type: application/json`.
servers:
  - url: /api/v1
security:
  - bearerKey: []
  - basicKey: []
  - session: []
components:
  securitySchemes:
    bearerKey: { type: http, scheme: bearer, description: "ACCESS_KEY_ID:SECRET" }
    basicKey: { type: http, scheme: basic }
    session: { type: apiKey, in: cookie, name: acs_session }
  parameters:
    bucket: { name: bucket, in: path, required: true, schema: { type: string } }
    key: { name: key, in: path, required: true, description: Object key (slashes allowed), schema: { type: string } }
...
    versionId: { name: versionId, in: query, schema: { type: string } }
...
paths:
...
  /buckets/{bucket}/objects:
    get:
      summary: List objects
      parameters:
        - { $ref: "#/components/parameters/bucket" }
        - { name: prefix, in: query, schema: { type: string } }
        - { name: delimiter, in: query, schema: { type: string }, description: 'Use "/" for folder-style listing' }
        - { name: cursor, in: query, schema: { type: string } }
        - { name: limit, in: query, schema: { type: integer, maximum: 1000 } }
      responses: { "200": { description: "{ objects, prefixes, nextCursor }" } }
  /buckets/{bucket}/objects/{key}:
    parameters: [{ $ref: "#/components/parameters/bucket" }, { $ref: "#/components/parameters/key" }]
    get:
      summary: Download (supports Range and conditional headers)
      parameters: [{ $ref: "#/components/parameters/versionId" }, { name: download, in: query, schema: { type: integer, enum: [1] } }]
      responses: { "200": { description: Object content }, "206": { description: Partial content } }
    put:
      summary: Upload with a raw request body (Content-Type is stored; X-Acs-Meta-* headers become metadata)
      requestBody: { content: { application/octet-stream: { schema: { type: string, format: binary } } } }
      responses: { "201": { description: Created } }
    delete:
      summary: Delete (creates a delete marker in versioned buckets, or removes versionId)
      parameters: [{ $ref: "#/components/parameters/versionId" }]
      responses: { "200": { description: "{ versionId, deleteMarker }" } }
...